User can play randomly a song from main activity
User is now capable of use a shortcut in main activity to select a song randomly and play it.

There have been some visual improvements in recycler views (main activity and show list option in play control activity) so when user reaches bounds the scrim spreads to the whole width of its container.