Player state saving added and fixed bug when searching a song.
Player now saves state again. It was removed due to some refactors and more implementation changes. Also it checks for inconsistency when loading state (e.g. when a song is removed from device or a playlist is deleted).

When searching and selecting a song, a different song was used to be set instead of the selected one. Now the app sets correctly the selected song.