Change style activity fragments now allows user to change UI elements colors properly.
When accessing change style activity, user can select the colors they want for a specific UI item. Changes are not saved and current style is not loaded when accessing change style activity.

Also, when rounded image layout is set for play control activity or playlists list fragment the inner circle gets the color of the background make items look like a CD.